Move LayoutNode and related traits to script_layout_interface.
