+++ components/script_layout_interface/Cargo.toml
@@ -11,8 +11,14 @@ path = "lib.rs"
 [dependencies]
 bitflags = "0.7"
 canvas_traits = {path = "../canvas_traits"}
+gfx_traits = {path = "../gfx_traits"}
 heapsize = "0.3.0"
 heapsize_plugin = "0.1.2"
 ipc-channel = {git = "https://github.com/servo/ipc-channel"}
+msg = {path = "../msg"}
 plugins = {path = "../plugins"}
+range = {path = "../range"}
+selectors = {version = "0.6", features = ["heap_size"]}
+string_cache = {version = "0.2.20", features = ["heap_size"]}
 style = {path = "../style"}
+url = {version = "1.0.0", features = ["heap_size"]}
